refactor various bounding box utilities
