+++ packages/tools/src/utilities/boundingBox/index.ts
@@ -1,0 +1,4 @@
+import extend2DBoundingBoxInViewAxis from './extend2DBoundingBoxInViewAxis';
+import getBoundingBoxAroundShape from './getBoundingBoxAroundShape';
+
+export { extend2DBoundingBoxInViewAxis, getBoundingBoxAroundShape };
