+++ packages/tools/src/utilities/segmentation/index.ts
@@ -1,7 +1,3 @@
-import {
-  getBoundingBoxAroundShape,
-  extend2DBoundingBoxInViewAxis,
-} from './getBoundingBoxUtils';
 import thresholdVolumeByRange from './thresholdVolumeByRange';
 import rectangleROIThresholdVolumeByRange from './rectangleROIThresholdVolumeByRange';
 import createMergedLabelmapForIndex from './createMergedLabelmapForIndex';
@@ -10,8 +6,6 @@ import getDefaultRepresentationConfig from './getDefaultRepresentationConfig';
 import createLabelmapVolumeForViewport from './createLabelmapVolumeForViewport';
 
 export {
-  getBoundingBoxAroundShape,
-  extend2DBoundingBoxInViewAxis,
   thresholdVolumeByRange,
   createMergedLabelmapForIndex,
   isValidRepresentationConfig,
