+++ packages/tools/src/utilities/segmentation/rectangleROIThresholdVolumeByRange.ts
@@ -1,4 +1,3 @@
-import { utilities as csUtils } from '@cornerstonejs/core';
 import type { Types } from '@cornerstonejs/core';
 import { state } from '../../stateManagement/annotation';
 import {
@@ -6,13 +5,8 @@ import {
   RectangleROIThresholdTool,
 } from '../../tools';
 
-import {
-  getBoundingBoxAroundShape,
-  extend2DBoundingBoxInViewAxis,
-} from '../segmentation';
 import thresholdVolumeByRange from './thresholdVolumeByRange';
-
-const { transformWorldToIndex } = csUtils;
+import getBoundsIJKFromRectangleROIs from '../boundingBox/getBoundsIJKFromRectangleROIs';
 
 export type ThresholdRangeOptions = {
   lower: number;
@@ -66,7 +60,7 @@ function rectangleROIThresholdVolumeByRange(
   const optionsToUse = {
     lower: options.lower,
     higher: options.higher,
-    boundsIJK: _getBoundsIJKFromRectangleROIs(
+    boundsIJK: getBoundsIJKFromRectangleROIs(
       annotations,
       referenceVolume,
       options
@@ -97,81 +91,6 @@ function _validateAnnotations(annotations) {
       );
     }
   }
-}
-
-function _getBoundsIJKFromRectangleROIs(annotations, referenceVolume, options) {
-  const { numSlicesToProject } = options;
-
-  const AllBoundsIJK = [];
-  annotations.forEach((annotation) => {
-    const { data } = annotation;
-    const { points } = data.handles;
-
-    const { imageData, dimensions } = referenceVolume;
-
-    let pointsToUse = points;
-    // If the tool is a 2D tool but has projection points, use them
-    if (data.cachedStats?.projectionPoints) {
-      const { projectionPoints } = data.cachedStats;
-      pointsToUse = [].concat(...projectionPoints); // cannot use flat() because of typescript compiler right now
-    }
-
-    const rectangleCornersIJK = pointsToUse.map(
-      (world) => transformWorldToIndex(imageData, world) as Types.Point3
-    );
-    let boundsIJK = getBoundingBoxAroundShape(rectangleCornersIJK, dimensions);
-
-    // If the tool is 2D but it is configured to project to X amount of slices
-    // Don't project the slices if projectionPoints have been used to define the extents
-    if (numSlicesToProject && !data.cachedStats?.projectionPoints) {
-      boundsIJK = extendBoundingBoxInSliceAxisIfNecessary(
-        boundsIJK,
-        numSlicesToProject
-      );
-    }
-
-    AllBoundsIJK.push(boundsIJK);
-  });
-
-  if (AllBoundsIJK.length === 1) {
-    return AllBoundsIJK[0];
-  }
-
-  // Get the intersection of all the bounding boxes
-  // This is the bounding box that contains all the ROIs
-  const boundsIJK = AllBoundsIJK.reduce(
-    (accumulator, currentValue) => {
-      return {
-        iMin: Math.min(accumulator.iMin, currentValue.iMin),
-        jMin: Math.min(accumulator.jMin, currentValue.jMin),
-        kMin: Math.min(accumulator.kMin, currentValue.kMin),
-        iMax: Math.max(accumulator.iMax, currentValue.iMax),
-        jMax: Math.max(accumulator.jMax, currentValue.jMax),
-        kMax: Math.max(accumulator.kMax, currentValue.kMax),
-      };
-    },
-    {
-      iMin: Infinity,
-      jMin: Infinity,
-      kMin: Infinity,
-      iMax: -Infinity,
-      jMax: -Infinity,
-      kMax: -Infinity,
-    }
-  );
-
-  return boundsIJK;
-}
-
-export function extendBoundingBoxInSliceAxisIfNecessary(
-  boundsIJK: [Types.Point2, Types.Point2, Types.Point2],
-  numSlicesToProject: number
-): [Types.Point2, Types.Point2, Types.Point2] {
-  const extendedBoundsIJK = extend2DBoundingBoxInViewAxis(
-    boundsIJK,
-    numSlicesToProject
-  );
-  return extendedBoundsIJK;
 }
 
 export default rectangleROIThresholdVolumeByRange;
